+++ src/tubes/matrix/Interpolation.java
@@ -10,7 +10,7 @@ public class Interpolation {
     for(int i = 0; i < matrix.getNRows(); i++) {
       for(int j = 0; j < matrix.getNCols(); j++) {
         if(j != matrix.getNCols() - 1) {
-          matrix.data[i][j] = (float) Math.pow(points.data[i][0], j);
+          matrix.data[i][j] = (double) Math.pow(points.data[i][0], j);
         } else {
           matrix.data[i][j] = points.data[i][1];
         }
@@ -20,12 +20,12 @@ public class Interpolation {
     return LinearEquation.cramerRule(matrix);
   }
 
-  public static float approximateFunction(Matrix points, float x) {
+  public static double approximateFunction(Matrix points, double x) {
     // Approximate the value of function at x by taking several points as input
     // and then approximate the function using polynomial interpolation method
 
     Matrix coefficients = polynomialInterpolation(points);
-    float output = 0.0f;
+    double output = 0.0f;
     for(int i = 0; i < coefficients.getNRows(); i++) {
       output += coefficients.data[i][0]*Math.pow(x, i);
     }
